requirements.txt
PyAudio~=0.2.14
vosk~=0.3.45
openai-whisper